01_original_study/code/requirements.txt
# ============================================================================
# APTOS 2019 Diabetic Retinopathy Classification - Python 3.10 Requirements
# Production-Ready Deep Learning Environment
# ============================================================================

# ============================================================================
# CORE DEEP LEARNING
# ============================================================================
# TensorFlow/Keras for CViTS-Net
tensorflow==2.15.0         # TensorFlow with Keras
tensorflow-addons==0.23.0  # Additional TensorFlow operations

# PyTorch (for other experiments)
torch==2.1.2               # PyTorch core framework
torchvision==0.16.2        # Computer vision utilities
torchaudio==2.1.2          # Audio processing (future use)
timm==0.9.12               # PyTorch Image Models (EfficientNet, ResNet)
tensorboard==2.15.1        # Training visualization
tensorboardX==2.6.2.2      # TensorBoard extensions

# ============================================================================
# DATA PROCESSING & ANALYSIS
# ============================================================================
pandas==2.1.4              # Data manipulation and CSV handling
numpy==1.24.3              # Numerical computing
scipy==1.11.4              # Scientific computing utilities
scikit-learn==1.3.2        # ML metrics (F1, confusion matrix, etc.)
imbalanced-learn==0.11.0   # Class imbalance handling (SMOTE, sampling)

# ============================================================================
# IMAGE PROCESSING & AUGMENTATION
# ============================================================================
opencv-python==4.8.1.78    # Image I/O and preprocessing
pillow==10.1.0             # Image manipulation
albumentations==1.3.1      # Medical imaging augmentation (MixUp, CutMix)

# ============================================================================
# VISUALIZATION & PLOTTING
# ============================================================================
matplotlib==3.8.2          # Publication-quality plotting
seaborn==0.13.0            # Statistical data visualization

# ============================================================================
# INTERACTIVE & NOTEBOOK ENVIRONMENTS
# ============================================================================
jupyter==1.0.0             # Jupyter Notebook base
jupyterlab==4.0.9          # JupyterLab IDE
notebook==7.0.6            # Classic notebook interface
ipykernel==6.26.0          # IPython kernel for notebooks

# ============================================================================
# UTILITIES & QUALITY OF LIFE
# ============================================================================
tqdm==4.66.1               # Progress bars for training loops
pyyaml==6.0.1              # YAML config file handling
requests==2.31.0           # HTTP requests (model downloads)
python-dotenv==1.0.0       # Environment variable management
colorama==0.4.6            # Colored terminal output

# ============================================================================
# OPTIONAL: DEVELOPMENT & DEBUGGING
# ============================================================================
# Uncomment if needed for development:
# ipython==8.17.2          # Enhanced interactive Python shell
# black==23.12.0           # Code formatter (PEP 8)
# flake8==6.1.0            # Code linter
# pytest==7.4.3            # Unit testing framework

# ============================================================================
# NOTES:
# ============================================================================
# Python Version: 3.10.x (Recommended: 3.10.13 or later)
# Environment: venv (d:\Ece_DR\DR_Research-main\aptos_env)
# CUDA: 12.1 (or 11.8 if older GPU drivers)
#
# Installation:
#   pip install -r requirements.txt
#
# Generate this file:
#   pip freeze > requirements.txt
#
# Verify installations:
#   python -c "import torch; print(torch.__version__, torch.cuda.is_available())"
# ============================================================================
